pyproject.toml
[tool.poetry]
name = "conversor temperatura"
version = "0.1.0"
description = "Conversor básico de temperatura usando Tkinter"
authors = ["MiguelAngelNarvaez <mignarvaez@umariana.edu.co>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
